Assorted lint tweaks
* use single quotes
* consistently use const/let
* use template literals
* fix no-promise-executor-return
* return early in a couple of cases
* use `for...of` when possible
* stop using implicit coercion for Strings
* use a ternary when possible
* rename variables to be descriptive
* use `new Error` when throwing an error
* use Object spread instead of Object.assign/Array.from
* prefer `append`/`remove`
* prefer `Set.has`
* use the strict comparison operators everywhere except for null
* use one-line expressions in some cases
* return early in a few cases
* use destructuring in more places
* remove a few functions used once
* other minor tweaks

+++ src/js/page/worker-messenger.js
@@ -19,7 +19,9 @@ export default class WorkerMessenger {
     return new Promise((resolve, reject) => {
       message.id = ++this._requestId;
       this._pending[message.id] = [resolve, reject];
-      this._postMessage(message);
+
+      if (!this._worker) this._startWorker();
+      this._worker.postMessage(message);
     });
   }
 
@@ -27,15 +29,13 @@ export default class WorkerMessenger {
     if (Object.keys(this._pending).length === 0) return;
 
     this._abortPending();
-    if (this._worker) {
-      this._worker.terminate();
-    }
+    if (this._worker) this._worker.terminate();
     this._startWorker();
   }
 
   _abortPending() {
     for (const key of Object.keys(this._pending)) {
-      this._fulfillPending(key, null, new DOMException("AbortError", "AbortError"));
+      this._fulfillPending(key, null, new DOMException('AbortError', 'AbortError'));
     }
   }
 
@@ -44,16 +44,9 @@ export default class WorkerMessenger {
     this._worker.onmessage = event => this._onMessage(event);
   }
 
-  _postMessage(message) {
-    if (!this._worker) {
-      this._startWorker();
-    }
-    this._worker.postMessage(message);
-  }
-
   _onMessage(event) {
     if (!event.data.id) {
-      console.log("Unexpected message", event);
+      console.log('Unexpected message', event);
       return;
     }
 
@@ -68,7 +61,7 @@ export default class WorkerMessenger {
     const resolver = this._pending[id];
 
     if (!resolver) {
-      console.log("No resolver for", { id, result, error });
+      console.log('No resolver for', { id, result, error });
       return;
     }
